Added support for 2d rendering by moving the projection matrix calculation to the camera. Added a new camera (Camera2D) so that ortho rendering can be done. To render with 2d instead of with 3d simply use setCamera(new Camera2D()) function in the constructor of your subclassed RajawaliRenderer.
Also added a new direction paramater to the Plane
class so that it can be created standing up on the y
axiss, this makes creating a 2d plane that is visible
by the 2d Camera by default easier.

By default the 2d camera renders the scene without
taking the aspect ratio into consideration, this is done
so that it is as easy as possible to get a plane
that covers the entire view (for live shader based wallpapers).