Fix Gouraud shader compilation w/ multiple lights
float NdotL was declared in a loop for every light. Multiple lights
caused the shader compilation to fail due to multiple declarations:

[rajawali.materials.GouraudMaterial] Could not compile vertex shader:
Shader log: (63) : error C1038: declaration of "NdotL" conflicts with
previous declaration at (57)

Declaration of NdotL was moved out of the loop and to the top of
main() so that it is only declared once.