Prevent object picker from failing after rotation
The object picker's buffers are now large enough to accommodate both
portrait and landscape orientations. Before the change, the picker
stopped after rotating the screen and logged the following error
message: "Could not bind FrameBuffer for color picking."

The downside is that more memory is being consumed. Ideally, the object
picker code should be updated to detect resolution changes and then
reinitialize or resize the buffers accordingly.