Add project-specific code formatting profile
The default Eclipse profile is too aggressive and globs line breaks
intentionally placed by the programmer. This profile is different from
the Eclipse default in the following ways:
* Max line width for code and comments is 120 (default was 80)
* Never joins already wrapped lines (preserves existing format and line
breaks in things like array initializers)
* Inserts blank line before first declaration in class declarations
* Does not insert blank line in empty bodies (e.g. `class Empty {}`)
* Does not put empty statement on new line (e.g. `while (something);`)
* Inserts new line at end of file
* Enables header comment formatting

These changes are up for debate, but I thought this would be a good
starting point.

Contributors are encouraged to apply formatting and functional changes
as separate commits. It is easier to review changes when they aren't
mixed.