Added the option to specify the export directory
Added method setExportDirectory(File exportDir) which allows the user to tell the exporter where to place files.
If no directory is specified, it defaults to the external storage directory, which it invariably wrote to before.
The user is responsible for the directory's validity. Any issues with the directory are caught in the existing methods.